pyproject.toml
[project]
name = "mcp-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.0",
    "mcp[cli]>=1.15.0",
    "nest-asyncio>=1.6.0",
    "pdfplumber>=0.11.9",
    "playwright>=1.55.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
